clean-up config and remove proxy functionality
