+++ config.js
@@ -8,12 +8,8 @@ const config = {
   user: null,
   cache: 0,
   limit: '50mb',
-  servers: {
-    web: {
-      host: '0.0.0.0',
-      port: 9999
-    }
-  },
+  host: '0.0.0.0',
+  port: 9999,
   data: path.join(__dirname, '/data'),
   routes: path.join(__dirname, '/routes.json')
