+++ server/middleware/tests/mulit-part-spec.js
@@ -12,6 +12,7 @@ describe('Multi-part', () => {
   helper.serverSpecHelper();
 
   it('should respond with dummy-response-1.json for empty form fields and one file attachment', (done) => {
+
     request.post(helper.getUrl('/v1/route5'))
       .type('multipart/form-data')
       .attach('attatchment', helper.getFile('dummy-response-1.json'))
@@ -21,9 +22,11 @@ describe('Multi-part', () => {
         expect(_.isEqual(res.body, {'a': 1})).to.be.ok;
         done();
       });
+
   });
 
   it('should respond with dummy-response-2.json for empty form fields and one file attachment', (done) => {
+
     request.post(helper.getUrl('/v1/route5'))
       .type('multipart/form-data')
       .attach('attachment', helper.getFile('dummy-response-2.json'))
@@ -33,9 +36,11 @@ describe('Multi-part', () => {
         expect(_.isEqual(res.body, {'b': 2})).to.be.ok;
         done();
       });
+
   });
 
   it('should respond with dummy-response-2.json for 1 matching form field and one file attachment', (done) => {
+
     request.post(helper.getUrl('/v1/route5'))
       .type('multipart/form-data')
       .attach('attachment', helper.getFile('dummy-response-2.json'))
@@ -46,9 +51,11 @@ describe('Multi-part', () => {
         expect(_.isEqual(res.body, {'b': 2})).to.be.ok;
         done();
       });
+
   });
 
   it('should respond with dummy-response-3.json for 2 matching form fields and one file attachment', (done) => {
+
     request.post(helper.getUrl('/v1/route5'))
       .type('multipart/form-data')
       .attach('attachment', helper.getFile('dummy-response-3.json'))
@@ -60,9 +67,11 @@ describe('Multi-part', () => {
         expect(_.isEqual(res.body, {'c': 3})).to.be.ok;
         done();
       });
+
   });
 
   it('should respond with dummy-response-4.json for field name that matches file input name', (done) => {
+
     request.post(helper.getUrl('/v1/route5'))
       .type('multipart/form-data')
       .attach('attachment', helper.getFile('dummy-response-4.json'))
@@ -74,5 +83,6 @@ describe('Multi-part', () => {
         expect(_.isEqual(res.body, {'d': 4})).to.be.ok;
         done();
       });
+
   });
 });
