+++ server/middleware/tests/request-spec.js
@@ -9,11 +9,13 @@ describe('Events', () => {
   helper.serverSpecHelper();
 
   it('should emit an event for all requests when a request is received', (done) => {
+
     helper.ekko.on('av:request', () => {
       done();
     });
 
     request.post(helper.getUrl('/v1/route9'))
     .end();
+
   });
 });
