+++ server/routes/tests/routes-spec.js
@@ -42,6 +42,7 @@ describe('Routes', () => {
   });
 
   it('route 1 should respond with dummy-response1.json', (done) => {
+
     request.get(helper.getUrl('/v1/route1'))
       .end((err, res) => {
         expect(err).to.be.null;
@@ -49,6 +50,7 @@ describe('Routes', () => {
         expect(_.isEqual(res.body, {'a': 1})).to.be.ok;
         done();
       });
+
   });
 
   it('route 2 should respond with dummy-response2.json for GET', (done) => {
@@ -60,6 +62,7 @@ describe('Routes', () => {
         expect(_.isEqual(res.body, {'b': 2})).to.be.ok;
         done();
       });
+
   });
 
   it('route 2 should respond with dummy-response3.json for POST', (done) => {
@@ -72,9 +75,11 @@ describe('Routes', () => {
         expect(_.isEqual(res.body, {'c': 3})).to.be.ok;
         done();
       });
+
   });
 
   it('route 4 should respond with dummy-response-2.json for POST with parameters', (done) => {
+
     request.post(helper.getUrl('/v1/route4'))
       .send({a: { b: 'b'}})
       .end((err, res) => {
@@ -83,6 +88,7 @@ describe('Routes', () => {
         expect(_.isEqual(res.body, {'b': 2})).to.be.ok;
         done();
       });
+
   });
 
   it('route 4 should response with dummy-response1.json [default file] for POST with no parameters', (done) => {
@@ -94,6 +100,7 @@ describe('Routes', () => {
         expect(_.isEqual(res.body, {'a': 1})).to.be.ok;
         done();
       });
+
   });
 
   it('route 9 should response with dummy-response1.json and status 201 for GET', (done) => {
@@ -116,28 +123,7 @@ describe('Routes', () => {
         expect(_.isEqual(res.body, {'b': 2})).to.be.ok;
         done();
       });
-  });
 
-  it('external mock 1 should response with dummy-response1.json due to override defined', (done) => {
-
-    request.post(helper.getUrl('/v1/external/route1'))
-      .end((err, res) => {
-        expect(err).to.be.null;
-        expect(res.status).to.equal(200);
-        expect(_.isEqual(res.body, {'a': 1})).to.be.ok;
-        done();
-      });
-  });
-
-  it('external mock 2 should response with external-dummy-response-1.json', (done) => {
-
-    request.post(helper.getUrl('/v1/external/route2'))
-      .end((err, res) => {
-        expect(err).to.be.null;
-        expect(res.status).to.equal(200);
-        expect(_.isEqual(res.body, {'z': 26})).to.be.ok;
-        done();
-      });
   });
 
 });
