--- server/tests/external-mock/data/external-dummy-response-1.json
@@ -1,1 +1,0 @@
-{"z": 26}
